separação de dom e firebase dos posts
