+++ src/style.css
@@ -6,7 +6,8 @@
 }
 
 body{
-    background-image: linear-gradient(#171923, #c31331);
+    background-image: url("img/fundo-celular.jpg");
+    background-size: cover;
 }
 
 .overlay {
@@ -21,7 +22,7 @@ body{
 
 .logo-img{
     display: flex;
-    width: 60%;
+    width: 50%;
     height: 90%;
     margin: 0 auto;
     align-items: center;
@@ -90,7 +91,7 @@ body{
     background-color: rgba(255, 255, 255, 0.589);
     box-shadow: 5px 10px 20px rgba(0, 0, 0, 0.651);
     width: 70%;
-    margin: 5;
+    margin: 10px;
     margin-left: 60px;
     border-radius: 20px;
 }
@@ -491,5 +492,5 @@ body{
 }
 
 .like:hover{
-    background-color: red;
+    background-color: yellow;
 }
